Make learning rate a model parameter
The default learning rate is set for volumetric
rendering.  Once we moved to tomographic rendering,
and no longer had the exponential attenuation,
the results show that a learning rate of 5e-4 is
too high, and results in training curves that are
very spiky.  Accordingly, we want to do a grid
search with learning rates orders of magnitude below
the volumetric rendering value.